Step5

+++ src/main/java/org/example/dto/ExchangeRateResponseDto.java
@@ -2,8 +2,9 @@ package org.example.dto;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Builder
-public record ExchangeRateResponseDto(String code, Long rate, LocalDate date) {
+public record ExchangeRateResponseDto(String code, BigDecimal rate, LocalDate date) {
 }
